Separate your account and manage users controller
This also meant creating a dedicate namespace for modules' submodules.
