+++ resources/views/template/navbar.blade.php
@@ -345,7 +345,7 @@
                             <!-- Dropdown - User Information -->
                             <div class="dropdown-menu dropdown-menu-right shadow animated--grow-in"
                                 aria-labelledby="userDropdown">
-                                <a class="dropdown-item" href="/account">
+                                <a class="dropdown-item" href="/account/me">
                                     <i class="fas fa-sign-out-alt fa-sm fa-fw mr-2 text-gray-400"></i>
                                     Account Settings
                                 </a>
